+++ js/Demo/demoMod.js
@@ -11,8 +11,8 @@ let modInfo = {
 
 // Set your version in num and name
 let VERSION = {
-	num: "2.5.11",
-	name: "Dreams Really Do Come True",
+	num: "2.6",
+	name: "Fixed Reality",
 }
 
 let changelog = `<h1>Changelog:</h1><br>