+++ js/technical/canvas.js
@@ -7,7 +7,6 @@ function retrieveCanvasData() {
 	let treeCanv = document.getElementById("treeCanvas")
 	let treeTab = document.getElementById("treeTab")
 	if (treeCanv===undefined||treeCanv===null) return false;
-	if (treeTab===undefined||treeTab===null) return false;
 	canvas = treeCanv;
 	ctx = canvas.getContext("2d");
 	return true;
@@ -46,13 +45,21 @@ function drawTree() {
 					drawTreeBranch(layer, tmp[layer].branches[branch])
 				}
 		}
+		for(id in layers[layer].upgrades) {
+			if (tmp[layer].upgrades[id].branches) {
+				for (branch in tmp[layer].upgrades[id].branches)
+				{
+					drawTreeBranch(id, tmp[layer].upgrades[id].branches[branch], "upgrade-" + layer + "-")
+				}
+
+			}
+		}
 	}
 }
 
-function drawTreeBranch(num1, data) { // taken from Antimatter Dimensions & adjusted slightly
+function drawTreeBranch(num1, data, prefix) { // taken from Antimatter Dimensions & adjusted slightly
 	let num2 = data
 	let color_id = 1
-
 	if (Array.isArray(data)){
 		num2 = data[0]
 		color_id = data[1]
@@ -60,7 +67,10 @@ function drawTreeBranch(num1, data) { // taken from Antimatter Dimensions & adju
 
 	if(typeof(color_id) == "number")
 		color_id = colors_theme[color_id]
-
+	if (prefix) {
+		num1 = prefix + num1
+		num2 = prefix + num2
+	}
 	if (document.getElementById(num1) == null || document.getElementById(num2) == null)
 		return