+++ js/utils.js
@@ -192,6 +192,7 @@ function buyUpg(layer, id) {
 	player[layer].upgrades.push(id);
 	if (upg.onPurchase != undefined)
 		run(upg.onPurchase, upg)
+	needCanvasUpdate = true
 	return true
 }
 
@@ -239,7 +240,8 @@ function inChallenge(layer, id) {
 	if (challenge == id) return true
 
 	if (layers[layer].challenges[challenge].countsAs)
-		return tmp[layer].challenges[challenge].countsAs.includes(id)
+		return tmp[layer].challenges[challenge].countsAs.includes(id) || false
+	return false
 }
 
 // ************ Misc ************
@@ -268,7 +270,7 @@ function showNavTab(name, prev) {
 	if (tmp[name] && tmp[name].previousTab !== undefined) prev = tmp[name].previousTab
 	var toTreeTab = name == "tree-tab"
 	console.log(name + prev)
-	if (!tmp[prev]?.leftTab == !tmp[name]?.leftTab) player[name].prevTab = prev
+	if (name!== "none" && prev && !tmp[prev]?.leftTab == !tmp[name]?.leftTab) player[name].prevTab = prev
 	else if (player[name])
 		player[name].prevTab = ""
 	player.navTab = name
@@ -281,7 +283,7 @@ function goBack(layer) {
 	let nextTab = "none"
 
 	if (player[layer].prevTab) nextTab = player[layer].prevTab
-	if (player.navTab !== "none" && (tmp[layer]?.row == "side" || tmp[layer].row == "side")) nextTab = player.lastSafeTab
+	if (player.navTab === "none" && (tmp[layer]?.row == "side" || tmp[layer].row == "otherside")) nextTab = player.lastSafeTab
 
 	if (tmp[layer].leftTab) showNavTab(nextTab, layer)
 	else showTab(nextTab, layer)
@@ -339,13 +341,7 @@ function subtabResetNotify(layer, family, id) {
 }
 
 function nodeShown(layer) {
-	if (layerShown(layer)) return true
-	switch (layer) {
-		case "idk":
-			return player.idk.unlocked
-			break;
-	}
-	return false
+	return layerShown(layer)
 }
 
 function layerunlocked(layer) {
@@ -368,6 +364,7 @@ function updateMilestones(layer) {
 	for (id in layers[layer].milestones) {
 		if (!(hasMilestone(layer, id)) && layers[layer].milestones[id].done()) {
 			player[layer].milestones.push(id)
+			if (layers[layer].milestones[id].onComplete) layers[layer].milestones[id].onComplete()
 			if (tmp[layer].milestonePopups || tmp[layer].milestonePopups === undefined) doPopup("milestone", tmp[layer].milestones[id].requirementDescription, "Milestone Gotten!", 3, tmp[layer].color);
 			player[layer].lastMilestone = id
 		}